+++ Makefile
@@ -1,7 +1,6 @@
 link:
 	cd hedvig-redux; yarn link
-	cd hedvig-style; yarn link
-	cd hedvig-app; yarn link hedvig-redux; yarn link hedvig-style
+	cd hedvig-app; yarn link hedvig-redux;
 
 install:
 	cd hedvig-redux; yarn install
