+++ hedvig-app/src/components/Dialog.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Dimensions, BackHandler, Text } from 'react-native';
 import PopupDialog from 'react-native-popup-dialog';
-import { theme } from 'hedvig-style';
+import { theme } from '../style-theme';
 import {
   DialogContainer,
   Heading,
