+++ hedvig-app/src/style-theme.js
