--- hedvig-style/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "hedvig-style",
-  "version": "1.0.0",
-  "description": "Common styling for hedvig-app and hedvig-web",
-  "main": "src/index.js",
-  "author": "Pascal Chatterjee & Fredrik Stockman",
-  "license": "AGPL",
-  "private": true,
-  "dependencies": {}
-}
